format fixes

+++ src/fs/ufs/store_io_ufs.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store_io_ufs.cc,v 1.7 2001/03/03 10:39:39 hno Exp $
+ * $Id: store_io_ufs.cc,v 1.8 2001/10/24 06:55:45 hno Exp $
  *
  * DEBUG: section 79    Storage Manager UFS Interface
  * AUTHOR: Duane Wessels
@@ -227,8 +227,8 @@ storeUfsWriteDone(int fd, int errflag, size_t len, void *my_data)
 {
     storeIOState *sio = my_data;
     ufsstate_t *ufsstate = (ufsstate_t *) sio->fsstate;
-    debug(79, 3) ("storeUfsWriteDone: dirno %d, fileno %08X, FD %d, len %d\n",
-	sio->swap_dirn, sio->swap_filen, fd, len);
+    debug(79, 3) ("storeUfsWriteDone: dirno %d, fileno %08X, FD %d, len %ld\n",
+	sio->swap_dirn, sio->swap_filen, fd, (long int)len);
     ufsstate->flags.writing = 0;
     if (errflag) {
 	debug(79, 0) ("storeUfsWriteDone: got failure (%d)\n", errflag);
